+++ fyp.sql
@@ -32,7 +32,7 @@ CREATE TABLE `attraction` (
   `image` varchar(255) DEFAULT NULL,
   `createTime` timestamp NULL DEFAULT CURRENT_TIMESTAMP,
   `editTime` date DEFAULT NULL,
-  PRIMARY KEY (`attactionID`)
+  PRIMARY KEY (`attractionID`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
 /*!40101 SET character_set_client = @saved_cs_client */;
